Merged changes from cad_audio:hifi_nnlib_latest_versions (commit id:c8737cb82e4d7758fadd5ff1dd6ef5bf3716125e).

+++ tensorflow/lite/micro/tools/make/ext_libs/xtensa_download.sh
@@ -47,13 +47,17 @@ if [ ! -d ${DOWNLOADS_DIR} ]; then
 fi
 
 if [[ ${2} == "hifi4" ]]; then
-  LIBRARY_URL="http://github.com/foss-xtensa/nnlib-hifi4/raw/master/archive/xa_nnlib_hifi4_08_02_2022.zip"
+  LIBRARY_URL="http://github.com/foss-xtensa/nnlib-hifi4/raw/master/archive/xa_nnlib_hifi4_08_12_2022.zip"
   LIBRARY_DIRNAME="xa_nnlib_hifi4"
-  LIBRARY_MD5="d930ec512df8e7dc814602319d2948b5"
+  LIBRARY_MD5="e4260f680716c5d6bd7ea958e3c40123"
 elif [[ ${2} == "hifi5" ]]; then
-  LIBRARY_URL="http://github.com/foss-xtensa/nnlib-hifi5/raw/master/archive/xa_nnlib_hifi5_02_01_2022.zip"
+  LIBRARY_URL="http://github.com/foss-xtensa/nnlib-hifi5/raw/master/archive/xa_nnlib_hifi5_08_01_2022.zip"
   LIBRARY_DIRNAME="xa_nnlib_hifi5"
-  LIBRARY_MD5="52b55bc4ad473eeebc3dc49792c527bd"
+  LIBRARY_MD5="826459ecc18c4089f8e705a66cb0e081"
+elif [[ ${2} == "vision_p6" ]]; then
+  LIBRARY_URL="https://github.com/foss-xtensa/tflmlib_vision/raw/main/archive/xi_tflmlib_vision_p6_22_06_29.zip"
+  LIBRARY_DIRNAME="xi_tflmlib_vision_p6"
+  LIBRARY_MD5="fea3720d76fdb3a5a337ace7b6081b56"
 else
   echo "Attempting to download an unsupported xtensa variant: ${2}"
   exit 1
@@ -75,17 +79,22 @@ else
     exit 1
   fi
 
-  unzip -qo "$TEMPFILE" -d ${DOWNLOADS_DIR} >&2
+  # Check if another make process has already extracted the downloaded files.
+  # If so, skip extracting and patching.
+  if [ -d ${LIBRARY_INSTALL_PATH} ]; then
+    echo >&2 "${LIBRARY_INSTALL_PATH} already exists, skipping the extraction."
+  else
+    unzip -qo "$TEMPFILE" -d ${DOWNLOADS_DIR} >&2
 
-  rm -rf "${TEMPDIR}"
+    rm -rf "${TEMPDIR}"
 
-  pushd "${LIBRARY_INSTALL_PATH}" > /dev/null
-  chmod -R +w ./
-  create_git_repo ./
-  if [[ ${2} == "hifi5" ]]; then
-    apply_patch_to_folder ./ "../../ext_libs/xa_nnlib_${2}.patch" "TFLM patch"
+    pushd "${LIBRARY_INSTALL_PATH}" > /dev/null
+    chmod -R +w ./
+    if [[ -f "../../ext_libs/xa_nnlib_${2}.patch" ]]; then
+      create_git_repo ./
+      apply_patch_to_folder ./ "../../ext_libs/xa_nnlib_${2}.patch" "TFLM patch"
+    fi
   fi
 fi
-
 
 echo "SUCCESS"
